Have seperate compilation unit for shape utilities. Improve shape testing example.
