+++ DDCore/src/ShapeTags.h
@@ -1,0 +1,44 @@
+//==========================================================================
+//  AIDA Detector description implementation 
+//--------------------------------------------------------------------------
+// Copyright (C) Organisation europeenne pour la Recherche nucleaire (CERN)
+// All rights reserved.
+//
+// For the licensing terms see $DD4hepINSTALL/LICENSE.
+// For the list of contributors see $DD4hepINSTALL/doc/CREDITS.
+//
+// Author     : M.Frank
+//
+//==========================================================================
+#ifndef DD4HEP_DDCORE_SHAPETAGS_H
+#define DD4HEP_DDCORE_SHAPETAGS_H
+
+#define SHAPELESS_TAG       "Assembly"
+#define BOX_TAG             "Box"
+#define HALFSPACE_TAG       "HalfSpace"
+#define POLYCONE_TAG        "Polycone"
+#define CONE_TAG            "Cone"
+#define TUBE_TAG            "Tube"
+#define CUTTUBE_TAG         "CutTube"
+#define CONESEGMENT_TAG     "ConeSegment"
+#define TRD1_TAG            "Trd1"
+#define TRD2_TAG            "Trd2"
+#define PARABOLOID_TAG      "Paraboloid"
+#define HYPERBOLOID_TAG     "Hyperboloid"
+#define ELLIPTICALTUBE_TAG  "EllipticalTube"
+#define SPHERE_TAG          "Sphere"
+#define TORUS_TAG           "Torus"
+#define TRAP_TAG            "Trap"
+#define POLYHEDRA_TAG       "Polyhedra"
+#define EXTRUDEDPOLYGON_TAG "ExtrudedPolygon"
+#define EIGHTPOINTSOLID_TAG "EightPointSolid"
+
+#define TRUNCATEDTUBE_TAG   "TruncatedTube"
+#define PSEUDOTRAP_TAG      "PseudoTrap"
+#define UNION_TAG           "Union"
+#define SUBTRACTION_TAG     "Subtraction"
+#define INTERSECTION_TAG    "Intersection"
+
+
+#endif // DD4HEP_DDCORE_SHAPETAGS_H
+
