+++ DDCore/src/Shapes.cpp
@@ -18,6 +18,7 @@
 #include "DD4hep/MatrixHelpers.h"
 #include "DD4hep/DD4hepUnits.h"
 #include "DD4hep/Printout.h"
+#include "ShapeTags.h"
 
 // C/C++ include files
 #include <stdexcept>
@@ -35,990 +36,10 @@ using namespace std;
 using namespace dd4hep;
 namespace units = dd4hep;
 
-#define TRUNCATEDTUBE_TAG "TruncatedTube"
-#define PSEUDOTRAP_TAG    "PseudoTrap"
-#define UNION_TAG         "Union"
-#define SUBTRACTION_TAG   "Subtraction"
-#define INTERSECTION_TAG  "Intersection"
-
-namespace dd4hep {
-  static bool check_shape_type(const Handle<TGeoShape>& solid, const TClass* cl)   {
-    return solid.isValid() && solid->IsA() == cl;
-  }
-
-  /// Type check of various shapes. Result like dynamic_cast. Compare with python's isinstance(obj,type)
-  template <typename SOLID> bool isInstance(const Handle<TGeoShape>& solid)   {
-    return check_shape_type(solid, SOLID::Object::Class());
-  }
-  template bool isInstance<Box>(const Handle<TGeoShape>& solid);
-  template bool isInstance<ShapelessSolid>(const Handle<TGeoShape>& solid);
-  template bool isInstance<HalfSpace>(const Handle<TGeoShape>& solid);
-  template bool isInstance<ConeSegment>(const Handle<TGeoShape>& solid);
-  template bool isInstance<CutTube>(const Handle<TGeoShape>& solid);
-  template bool isInstance<EllipticalTube>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Trap>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Trd1>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Trd2>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Torus>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Sphere>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Paraboloid>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Hyperboloid>(const Handle<TGeoShape>& solid);
-  template bool isInstance<PolyhedraRegular>(const Handle<TGeoShape>& solid);
-  template bool isInstance<Polyhedra>(const Handle<TGeoShape>& solid);
-  template bool isInstance<ExtrudedPolygon>(const Handle<TGeoShape>& solid);
-  template bool isInstance<BooleanSolid>(const Handle<TGeoShape>& solid);
-
-  template <> bool isInstance<Cone>(const Handle<TGeoShape>& solid)  {
-    return check_shape_type(solid, TGeoConeSeg::Class())
-      ||   check_shape_type(solid, TGeoCone::Class());
-  }
-  template <> bool isInstance<Tube>(const Handle<TGeoShape>& solid)  {
-    return check_shape_type(solid, TGeoTubeSeg::Class())
-      ||   check_shape_type(solid, TGeoCtub::Class());
-  }
-  template <> bool isInstance<Polycone>(const Handle<TGeoShape>& solid)   {
-    return check_shape_type(solid, TGeoPcon::Class())
-      ||   check_shape_type(solid, TGeoPgon::Class());
-  }
-  template <> bool isInstance<EightPointSolid>(const Handle<TGeoShape>& solid)   {
-    if ( solid.isValid() )   {
-      TClass* c = solid->IsA();
-      return c==TGeoArb8::Class() || c==TGeoTrap::Class() || c==TGeoGtra::Class();
-    }
-    return false;
-  }
-  template <> bool isInstance<TruncatedTube>(const Handle<TGeoShape>& solid)   {
-    if ( solid.isValid() )   {
-      return solid->IsA() == TGeoCompositeShape::Class()
-        &&   ::strcmp(solid->GetTitle(), TRUNCATEDTUBE_TAG) == 0;
-    }
-    return false;
-  }
-  template <> bool isInstance<PseudoTrap>(const Handle<TGeoShape>& solid)   {
-    if ( solid.isValid() )   {
-      return solid->IsA() == TGeoCompositeShape::Class()
-        &&   ::strcmp(solid->GetTitle(), PSEUDOTRAP_TAG) == 0;
-    }
-    return false;
-  }
-  template <> bool isInstance<SubtractionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoSubtraction;
-  }
-  template <> bool isInstance<UnionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoUnion;
-  }
-  template <> bool isInstance<IntersectionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoIntersection;
-  }
-
-  /// Type check of various shapes. Do not allow for polymorphism. Types must match exactly
-  template <typename SOLID> bool isA(const Handle<TGeoShape>& solid)   {
-    return check_shape_type(solid, SOLID::Object::Class());
-  }
-  template bool isA<Box>(const Handle<TGeoShape>& solid);
-  template bool isA<ShapelessSolid>(const Handle<TGeoShape>& solid);
-  template bool isA<HalfSpace>(const Handle<TGeoShape>& solid);
-  template bool isA<Cone>(const Handle<TGeoShape>& solid);
-  template bool isA<ConeSegment>(const Handle<TGeoShape>& solid);
-  template bool isA<Tube>(const Handle<TGeoShape>& solid);
-  template bool isA<CutTube>(const Handle<TGeoShape>& solid);
-  template bool isA<EllipticalTube>(const Handle<TGeoShape>& solid);
-  template bool isA<Trap>(const Handle<TGeoShape>& solid);
-  template bool isA<Trd1>(const Handle<TGeoShape>& solid);
-  template bool isA<Trd2>(const Handle<TGeoShape>& solid);
-  template bool isA<Torus>(const Handle<TGeoShape>& solid);
-  template bool isA<Sphere>(const Handle<TGeoShape>& solid);
-  template bool isA<Paraboloid>(const Handle<TGeoShape>& solid);
-  template bool isA<Hyperboloid>(const Handle<TGeoShape>& solid);
-  template bool isA<PolyhedraRegular>(const Handle<TGeoShape>& solid);
-  template bool isA<Polyhedra>(const Handle<TGeoShape>& solid);
-  template bool isA<ExtrudedPolygon>(const Handle<TGeoShape>& solid);
-  template bool isA<Polycone>(const Handle<TGeoShape>& solid);
-  template bool isA<EightPointSolid>(const Handle<TGeoShape>& solid);
-
-  template <> bool isA<TruncatedTube>(const Handle<TGeoShape>& solid)   {
-    if ( solid.isValid() )   {
-      return solid->IsA() == TGeoCompositeShape::Class()
-        &&   ::strcmp(solid->GetTitle(), TRUNCATEDTUBE_TAG) == 0;
-    }
-    return false;
-  }
-  template <> bool isA<PseudoTrap>(const Handle<TGeoShape>& solid)   {
-    if ( solid.isValid() )   {
-      return solid->IsA() == TGeoCompositeShape::Class()
-        &&   ::strcmp(solid->GetTitle(), PSEUDOTRAP_TAG) == 0;
-    }
-    return false;
-  }
-  template <> bool isA<SubtractionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoSubtraction
-      &&    ::strcmp(solid->GetTitle(), SUBTRACTION_TAG) == 0;
-  }
-  template <> bool isA<UnionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoUnion
-      &&    ::strcmp(solid->GetTitle(), UNION_TAG) == 0;
-  }
-  template <> bool isA<IntersectionSolid>(const Handle<TGeoShape>& solid)   {
-    TGeoCompositeShape* sh = (TGeoCompositeShape*)solid.ptr();
-    return sh && sh->IsA() == TGeoCompositeShape::Class()
-      &&   sh->GetBoolNode()->GetBooleanOperator() == TGeoBoolNode::kGeoIntersection
-      &&    ::strcmp(solid->GetTitle(), INTERSECTION_TAG) == 0;
-  }
-}
-
-namespace dd4hep {
-  template <typename SHAPE> void invalidSetDimensionCall(const SHAPE& sh, const vector<double>& params)   {
-    stringstream str;
-    str << "Shape: " << typeName(typeid(sh)) << "::setDimension: "
-        << "Invalid number of parameters: " << params.size();
-    throw runtime_error(str.str());
-  }
-
-  /// Access Shape dimension parameters (As in TGeo, but angles in radians rather than degrees)
-  vector<double> get_shape_dimensions(TGeoShape* shape)   {
-    if (shape) {
-      TClass* cl = shape->IsA();
-      if (cl == TGeoShapeAssembly::Class()) {
-        const TGeoShapeAssembly* sh = (const TGeoShapeAssembly*) shape;
-        return { sh->GetDX(), sh->GetDY(), sh->GetDZ() };
-      }
-      else if (cl == TGeoBBox::Class()) {
-        const TGeoBBox* sh = (const TGeoBBox*) shape;
-        return { sh->GetDX(), sh->GetDY(), sh->GetDZ() };
-      }
-      else if (cl == TGeoHalfSpace::Class()) {
-        TGeoHalfSpace* sh = (TGeoHalfSpace*)(const_cast<TGeoShape*>(shape));
-        return { sh->GetPoint()[0], sh->GetPoint()[1], sh->GetPoint()[2],
-            sh->GetNorm()[0], sh->GetNorm()[1], sh->GetNorm()[2] };
-      }
-      else if (cl == TGeoPcon::Class()) {
-        const TGeoPcon* sh = (const TGeoPcon*) shape;
-        vector<double> pars { sh->GetPhi1()*units::deg, sh->GetDphi()*units::deg, double(sh->GetNz()) };
-        pars.reserve(3+3*sh->GetNz());
-        for (Int_t i = 0; i < sh->GetNz(); ++i) {
-          pars.emplace_back(sh->GetZ(i));
-          pars.emplace_back(sh->GetRmin(i));
-          pars.emplace_back(sh->GetRmax(i));
-        }
-        return pars;
-      }
-      else if (cl == TGeoConeSeg::Class()) {
-        const TGeoConeSeg* sh = (const TGeoConeSeg*) shape;
-        return { sh->GetDz(), sh->GetRmin1(), sh->GetRmax1(), sh->GetRmin2(), sh->GetRmax2(),
-            sh->GetPhi1()*units::deg, sh->GetPhi2()*units::deg };
-      }
-      else if (cl == TGeoCone::Class()) {
-        const TGeoCone* sh = (const TGeoCone*) shape;
-        return { sh->GetDz(), sh->GetRmin1(), sh->GetRmax1(), sh->GetRmin2(), sh->GetRmax2() };
-      }
-      else if (cl == TGeoTube::Class()) {
-        const TGeoTube* sh = (const TGeoTube*) shape;
-        return { sh->GetRmin(), sh->GetRmax(), sh->GetDz() };
-      }
-      else if (cl == TGeoTubeSeg::Class()) {
-        const TGeoTubeSeg* sh = (const TGeoTubeSeg*) shape;
-        return { sh->GetRmin(), sh->GetRmax(), sh->GetDz(), sh->GetPhi1()*units::deg, sh->GetPhi2()*units::deg };
-      }
-      else if (cl == TGeoCtub::Class()) {
-        const TGeoCtub* sh = (const TGeoCtub*) shape;
-        const Double_t*	lo = sh->GetNlow();
-        const Double_t*	hi = sh->GetNhigh();
-        return { sh->GetRmin(), sh->GetRmax(), sh->GetDz(),
-            sh->GetPhi1()*units::deg, sh->GetPhi2()*units::deg,
-            lo[0], lo[1], lo[2], hi[0], hi[1], hi[2] };
-      }
-      else if (cl == TGeoEltu::Class()) {
-        const TGeoEltu* sh = (const TGeoEltu*) shape;
-        return { sh->GetA(), sh->GetB(), sh->GetDz() };
-      }
-      else if (cl == TGeoTrd1::Class()) {
-        const TGeoTrd1* sh = (const TGeoTrd1*) shape;
-        return { sh->GetDx1(), sh->GetDx2(), sh->GetDy(), sh->GetDz() };
-      }
-      else if (cl == TGeoTrd2::Class()) {
-        const TGeoTrd2* sh = (const TGeoTrd2*) shape;
-        return { sh->GetDx1(), sh->GetDx2(), sh->GetDy1(), sh->GetDy2(), sh->GetDz() };
-      }
-      else if (cl == TGeoParaboloid::Class()) {
-        const TGeoParaboloid* sh = (const TGeoParaboloid*) shape;
-        return { sh->GetRlo(), sh->GetRhi(), sh->GetDz() };
-      }
-      else if (cl == TGeoHype::Class()) {
-        const TGeoHype* sh = (const TGeoHype*) shape;
-        return { sh->GetDz(),
-            sh->GetRmin(), sh->GetStIn()*units::deg,
-            sh->GetRmax(), sh->GetStOut()*units::deg };
-      }
-      else if (cl == TGeoSphere::Class()) {
-        const TGeoSphere* sh = (const TGeoSphere*) shape;
-        return { sh->GetRmin(), sh->GetRmax(),
-            sh->GetTheta1()*units::deg, sh->GetTheta2()*units::deg,
-            sh->GetPhi1()*units::deg,   sh->GetPhi2()*units::deg };
-      }
-      else if (cl == TGeoTorus::Class()) {
-        const TGeoTorus* sh = (const TGeoTorus*) shape;
-        return { sh->GetR(), sh->GetRmin(), sh->GetRmax(),
-            sh->GetPhi1()*units::deg, sh->GetDphi()*units::deg };
-      }
-      else if (cl == TGeoTrap::Class()) {
-        const TGeoTrap* sh = (const TGeoTrap*) shape;
-        return { sh->GetDz(), sh->GetTheta()*units::deg, sh->GetPhi()*units::deg,
-            sh->GetH1(), sh->GetBl1(), sh->GetTl1(), sh->GetAlpha1()*units::deg,
-            sh->GetH2(), sh->GetBl2(), sh->GetTl2(), sh->GetAlpha2()*units::deg };
-      }
-      else if (cl == TGeoPgon::Class()) {
-        const TGeoPgon* sh = (const TGeoPgon*) shape;
-        vector<double> pars { sh->GetPhi1()*units::deg, sh->GetDphi()*units::deg, double(sh->GetNedges()), double(sh->GetNz()) };
-        pars.reserve(4+3*sh->GetNz());
-        for (Int_t i = 0; i < sh->GetNz(); ++i) {
-          pars.emplace_back(sh->GetZ(i));
-          pars.emplace_back(sh->GetRmin(i));
-          pars.emplace_back(sh->GetRmax(i));
-        }
-        return pars;
-      }
-      else if (cl == TGeoXtru::Class()) {
-        const TGeoXtru* sh = (const TGeoXtru*) shape;
-        Int_t nz = sh->GetNz();
-        vector<double> pars { double(nz) };
-        pars.reserve(1+4*nz);
-        for(Int_t i=0; i<nz; ++i)   {
-          pars.emplace_back(sh->GetZ(i));
-          pars.emplace_back(sh->GetXOffset(i));
-          pars.emplace_back(sh->GetYOffset(i));
-          pars.emplace_back(sh->GetScale(i));
-        }
-        return pars;
-      }
-      else if (cl == TGeoArb8::Class())  {
-        TGeoTrap* sh = (TGeoTrap*) shape;
-        struct _V { double xy[8][2]; } *vtx = (_V*)sh->GetVertices();
-        vector<double> pars { sh->GetDz() };
-        for ( size_t i=0; i<8; ++i )   {
-          pars.emplace_back(vtx->xy[i][0]);
-          pars.emplace_back(vtx->xy[i][1]);
-        }
-        return pars;
-      }
-      else if (cl == TGeoScaledShape::Class()) {
-        TGeoScaledShape* sh = (TGeoScaledShape*) shape;
-        TGeoShape*       s_sh = sh->GetShape();
-        const Double_t*  scale = sh->GetScale()->GetScale();
-        vector<double>   pars {scale[0],scale[1],scale[2]};
-        vector<double>   s_pars = get_shape_dimensions(s_sh);
-        for(auto p : s_pars) pars.push_back(p);
-        return pars;
-      }
-      else if (cl == TGeoCompositeShape::Class()) {
-        Solid solid(shape);
-        const TGeoCompositeShape* sh = (const TGeoCompositeShape*) shape;
-        const TGeoBoolNode* boolean = sh->GetBoolNode();
-        TGeoMatrix* right_matrix  = boolean->GetRightMatrix();
-        TGeoShape*  left_solid    = boolean->GetLeftShape();
-        TGeoShape*  right_solid   = boolean->GetRightShape();
-        if ( isInstance<TruncatedTube>(solid) )   {
-          stringstream params(right_matrix->GetTitle());
-          vector<double> pars;
-          pars.reserve(8);
-          for(size_t i=0; i<8; ++i)   {
-            double val;
-            params >> val;
-            pars.emplace_back(val);
-            if ( !params.good() )   {
-              except("Solid::dimensions","+++ Invalid parameter stream.");
-            }
-          }
-          return pars;
-#if 0
-          TGeoTubeSeg* tubs = (TGeoTubeSeg*)left_solid;
-          TGeoBBox*    box  = (TGeoBBox*)right_solid;
-          double rmin       = tubs->GetRmin();
-          double rmax       = tubs->GetRmax();
-          double zhalf      = tubs->GetDz();
-          double startPhi   = tubs->GetPhi1()*units::degree;
-          double deltaPhi   = tubs->GetPhi2()*units::degree - startPhi;
-          double boxY       = box->GetDY();
-          double xBox       = right_matrix->GetTranslation()[0];
-          XYZAngles angles  = detail::matrix::_xyzAngles(right_matrix);
-          double alpha      = -angles.Z();
-          double sin_alpha  = std::sin(std::fabs(alpha));
-          double sin2alpha  = sin_alpha*sin_alpha;
-          double cos_phi    = std::cos(deltaPhi);
-          double cos2phi    = cos_phi*cos_phi;
-          double sin2phi    = 1.0 - cos2phi;
-          double ss1 = 1.0 - (sin2phi/sin2alpha);
-          double R = 0, r = 0;
-          bool   cutInside = true;
-
-          /// Need to select physical solutions: THIS DOES NOT WORK. CANNOT RESOLVE THE AMBIGUITY.
-          double r1 = xBox + boxY/sin_alpha;
-          double r2 = xBox - boxY/sin_alpha;
-          if ( r1 > 0 )
-            r = r1, cutInside = true;
-          else if ( r2 > 0 )
-            r = r2, cutInside = false;
-          else
-            except("","+++ Bad parameters");
-          
-          double R1 = -r*cos_phi/ss1*(1.0 + std::sqrt(1.0 + (1.0 - ss1/cos2phi)));
-          double R2 = -r*cos_phi/ss1*(1.0 - std::sqrt(1.0 + (1.0 - ss1/cos2phi)));
-          /// Need to select physical solutions: THIS DOES NOT WORK. CANNOT RESOLVE THE AMBIGUITY.
-          if ( R1 > 0 )
-            R = R1;
-          else if ( R2 > 0 )
-            R = R2;
-          else
-            except("","+++ Bad parameters");
-          double cutAtStart = r;
-          double cutAtDelta = R;
-
-          cout << "dimensions: " << endl;
-          cout << right_matrix->GetTitle() << endl;
-          cout << TRUNCATEDTUBE_TAG << ":" << endl
-               << "\t zhalf:     " << zhalf << " " << endl
-               << "\t rmin:      " << rmin << " " << endl
-               << "\t rmax:      " << rmax << " " << endl
-               << "\t startPhi:  " << startPhi/units::degree << " " << endl
-               << "\t deltaPhi:  " << deltaPhi/units::degree << " " << endl
-               << "\t cutAtStart:" << cutAtStart << " " << endl
-               << "\t cutAtDelta:" << cutAtDelta << " " << endl
-               << "\t cutInside: " << char(cutInside ? '1' : '0') << endl
-               << "\t\t boxY:      " << boxY << endl
-               << "\t\t alpha:     " << angles.Z() << " [rad] " << angles.Z()/units::degree << " [degree]" << endl
-               << "\t\t sin_alpha: " << sin_alpha << endl
-               << "\t\t xBox:      " << xBox << endl
-               << "\t\t r1:        " << r1 << endl
-               << "\t\t r2:        " << r2 << endl
-               << "\t\t R1:        " << R1 << endl
-               << "\t\t R2:        " << R2 << endl
-            ;
-
-          return {
-            zhalf, rmin, rmax,
-              startPhi*units::deg, deltaPhi*units::deg,
-              cutAtStart, cutAtDelta, cutInside ? 1.0 : 0.0 };
-#endif
-        }
-        else if ( isInstance<PseudoTrap>(solid) )   {
-          stringstream params(right_matrix->GetTitle());
-          vector<double> pars;
-          pars.reserve(7);
-#ifdef DIMENSION_DEBUG
-          cout << "dimensions: [" << PSEUDOTRAP_TAG << "]" << endl
-               << right_matrix->GetTitle() << endl;
-#endif
-          for(size_t i=0; i<7; ++i)   {
-            double val;
-            params >> val;
-            pars.emplace_back(val);
-            if ( !params.good() )   {
-              except("Solid::dimensions","+++ Invalid parameter stream.");
-            }
-          }
-          return pars;
-        }
-        else if ( isInstance<SubtractionSolid>(solid) )   {
-        }
-        else if ( isInstance<UnionSolid>(solid) )   {
-        }
-        else if ( isInstance<IntersectionSolid>(solid) )   {
-        }
-        
-        TGeoBoolNode::EGeoBoolType oper = boolean->GetBooleanOperator();
-        TGeoMatrix* left_matrix   = boolean->GetRightMatrix();
-        const Double_t* left_tr   = left_matrix->GetTranslation();
-        const Double_t* left_rot  = left_matrix->GetRotationMatrix();
-        const Double_t* right_tr  = right_matrix->GetTranslation();
-        const Double_t* right_rot = right_matrix->GetRotationMatrix();
-
-        vector<double> pars { double(oper) };
-        vector<double> left_par  = Solid(left_solid).dimensions();
-        vector<double> right_par = Solid(right_solid).dimensions();
-
-        pars.insert(pars.end(), left_par.begin(), left_par.end());
-        pars.insert(pars.end(), left_rot, left_rot+9);
-        pars.insert(pars.end(), left_tr, left_tr+3);
-
-        pars.insert(pars.end(), right_par.begin(), right_par.end());
-        pars.insert(pars.end(), right_rot, right_rot+9);
-        pars.insert(pars.end(), right_tr, right_tr+3);
-        return pars;
-      }
-      else  {
-        printout(ERROR,"Solid","Failed to access dimensions for shape of type:%s.",
-                 cl->GetName());
-      }
-      return {};
-    }
-    except("Solid","Failed to access dimensions [Invalid handle].");
-   return {};
-  }
-
-  /// Access Shape dimension parameters (As in TGeo, but angles in radians rather than degrees)
-  vector<double> get_shape_dimensions(Solid solid)   {
-    return get_shape_dimensions(solid.ptr());
-  }
-  
-  /// Set the shape dimensions. As for the TGeo shape, but angles in rad rather than degrees.
-  void set_shape_dimensions(TGeoShape* shape, const vector<double>& params)   {
-    if (shape) {
-      TClass* cl = shape->IsA();
-      Solid solid(shape);
-      if (cl == TGeoShapeAssembly::Class()) {
-        //TGeoShapeAssembly* sh = (TGeoShapeAssembly*) shape;
-        printout(WARNING,"ShapelessSolid","+++ setDimensions is a compatibility call. Nothing implemented.");
-      }
-      else if (cl == TGeoBBox::Class()) {
-        TGeoBBox* sh = (TGeoBBox*) shape;
-        auto pars = params;
-        if ( params.size() != 3 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoHalfSpace::Class()) {
-        TGeoHalfSpace* sh = (TGeoHalfSpace*)shape;
-        auto pars = params;
-        if ( params.size() != 6 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoPcon::Class()) {
-        TGeoPcon* sh = (TGeoPcon*) shape;
-        if ( params.size() < 3 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        size_t nz = size_t(params[2]);
-        if ( params.size() != 3 + 3*nz )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        vector<double> pars = params;
-        pars[0] /= units::deg;
-        pars[1] /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoConeSeg::Class()) {
-        TGeoConeSeg* sh = (TGeoConeSeg*) shape;
-        if ( params.size() != 7 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        auto pars = params;
-        pars[5] /= units::deg;
-        pars[6] /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoCone::Class()) {
-        TGeoCone* sh = (TGeoCone*) shape;
-        auto pars = params;
-        if ( params.size() != 5 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoTube::Class()) {
-        TGeoTube* sh = (TGeoTube*) shape;
-        if ( params.size() != 3 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        auto pars = params;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoTubeSeg::Class()) {
-        TGeoTubeSeg* sh = (TGeoTubeSeg*) shape;
-        if ( params.size() != 5 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        auto pars = params;
-        pars[3] /= units::deg;
-        pars[4] /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoCtub::Class()) {
-        TGeoCtub* sh = (TGeoCtub*) shape;
-        if ( params.size() != 11 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        auto pars = params;
-        pars[3] /= units::deg;
-        pars[4] /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoEltu::Class()) {
-        TGeoEltu* sh = (TGeoEltu*) shape;
-        auto pars = params;
-        if ( params.size() != 3 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoTrd1::Class()) {
-        TGeoTrd1* sh = (TGeoTrd1*) shape;
-        auto pars = params;
-        if ( params.size() != 4 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoTrd2::Class()) {
-        TGeoTrd2* sh = (TGeoTrd2*) shape;
-        auto pars = params;
-        if ( params.size() != 5 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoParaboloid::Class()) {
-        TGeoParaboloid* sh = (TGeoParaboloid*) shape;
-        auto pars = params;
-        if ( params.size() != 3 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoHype::Class()) {
-        TGeoHype* sh = (TGeoHype*) shape;
-        if ( params.size() != 5 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        double pars[] = { params[0], params[1], params[2]/units::deg, params[3], params[4]/units::deg };
-        solid._setDimensions(pars);
-      }
-      else if (cl == TGeoSphere::Class()) {
-        TGeoSphere* sh = (TGeoSphere*) shape;
-        if ( params.size() < 2 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        double pars[] = { params[0], params[1], 0.0, M_PI/units::deg, 0.0, 2*M_PI/units::deg };
-        if (params.size() > 2) pars[2] = params[2]/units::deg;
-        if (params.size() > 3) pars[3] = params[3]/units::deg;
-        if (params.size() > 4) pars[4] = params[4]/units::deg;
-        if (params.size() > 5) pars[5] = params[5]/units::deg;
-        sh->SetDimensions(pars, params.size());
-        sh->ComputeBBox();
-      }
-      else if (cl == TGeoTorus::Class()) {
-        TGeoTorus* sh = (TGeoTorus*) shape;
-        if ( params.size() != 5 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        double pars[] = { params[0], params[1], params[2], params[3]/units::deg, params[4]/units::deg };
-        solid._setDimensions(pars);
-      }
-      else if (cl == TGeoTrap::Class()) {
-        TGeoTrap* sh = (TGeoTrap*) shape;
-        auto pars = params;
-        if ( params.size() != 11 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        pars[1]  /= units::deg;
-        pars[2]  /= units::deg;
-        pars[6]  /= units::deg;
-        pars[10] /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoPgon::Class()) {
-        TGeoPgon* sh = (TGeoPgon*) shape;
-        auto pars = params;
-        if ( params.size() != 3 + 3*size_t(params[2]) )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        pars[0]  /= units::deg;
-        pars[1]  /= units::deg;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoXtru::Class()) {
-        TGeoXtru* sh = (TGeoXtru*) shape;
-        auto pars = params;
-        if ( params.size() != 1 + 4*size_t(params[0]) )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoArb8::Class())  {
-        TGeoTrap* sh = (TGeoTrap*) shape;
-        if ( params.size() != 17 )   {
-          invalidSetDimensionCall(*sh,params);
-        }
-        auto pars = params;
-        solid._setDimensions(&pars[0]);
-      }
-      else if (cl == TGeoScaledShape::Class()) {
-        TGeoScaledShape* sh = (TGeoScaledShape*) shape;
-        Solid            s_sh(sh->GetShape());
-        sh->GetScale()->SetScale(params[0], params[1], params[2]);
-        auto pars = params;
-        s_sh._setDimensions(&pars[3]);
-      }
-      else if (cl == TGeoCompositeShape::Class()) {
-        TGeoCompositeShape* sh = (TGeoCompositeShape*) shape;
-        TGeoBoolNode* boolean = sh->GetBoolNode();
-        TGeoMatrix* right_matrix  = boolean->GetRightMatrix();
-        TGeoShape*  left_solid    = boolean->GetLeftShape();
-        TGeoShape*  right_solid   = boolean->GetRightShape();
-        if ( isInstance<TruncatedTube>(solid) )   {
-          TGeoTubeSeg* tubs = (TGeoTubeSeg*)left_solid;
-          TGeoBBox*    box  = (TGeoBBox*)right_solid;
-          double zhalf = params[0];
-          double rmin  = params[1];
-          double rmax  = params[2];
-          double startPhi = params[3]/units::deg;
-          double deltaPhi = params[4]/units::deg;
-          double cutAtStart = params[5];
-          double cutAtDelta = params[6];
-          bool   cutInside  = params[7] > 0.5;
-#ifdef DIMENSION_DEBUG
-          cout << "setDimensions: [" << TRUNCATEDTUBE_TAG << "]" << endl
-               << right_matrix->GetTitle() << endl;
-#endif
-          // check the parameters
-          if( rmin <= 0 || rmax <= 0 || cutAtStart <= 0 || cutAtDelta <= 0 )
-            except(TRUNCATEDTUBE_TAG,"++ 0 <= rIn,cutAtStart,rOut,cutAtDelta,rOut violated!");
-          else if( rmin >= rmax )
-            except(TRUNCATEDTUBE_TAG,"++ rIn<rOut violated!");
-          else if( startPhi != 0. )
-            except(TRUNCATEDTUBE_TAG,"++ startPhi != 0 not supported!");
-
-          double r         = cutAtStart;
-          double R         = cutAtDelta;
-          // angle of the box w.r.t. tubs
-          double cath      = r - R * std::cos( deltaPhi*units::deg );
-          double hypo      = std::sqrt( r*r + R*R - 2.*r*R * std::cos( deltaPhi*units::deg ));
-          double cos_alpha = cath / hypo;
-          double alpha     = std::acos( cos_alpha );
-          double sin_alpha = std::sin( std::fabs(alpha) );
-  
-          // exaggerate dimensions - does not matter, it's subtracted!
-          // If we don't, the **edge** of the box would cut into the tube segment
-          // for larger delta-phi values
-          double boxX      = 1.1*rmax + rmax/sin_alpha; // Need to adjust for move!
-          double boxY      = rmax;
-          // width of the box > width of the tubs
-          double boxZ      = 1.1 * zhalf;
-          double xBox;      // center point of the box
-          if( cutInside )
-            xBox = r - boxY / sin_alpha;
-          else
-            xBox = r + boxY / sin_alpha;
-
-          // rotationmatrix of box w.r.t. tubs
-          TGeoRotation rot;
-          rot.RotateZ( -alpha/dd4hep::deg );
-          double box_dim[] = {boxX, boxY, boxZ};
-          double tub_dim[] = {rmin, rmax, zhalf, startPhi, deltaPhi};
-          box->SetDimensions(box_dim);
-          tubs->SetDimensions(tub_dim);
-          TGeoCombiTrans* combi = (TGeoCombiTrans*)right_matrix;
-          combi->SetRotation(rot);
-          combi->SetTranslation(xBox, 0, 0);
-          return;
-        }
-        else if ( isInstance<PseudoTrap>(solid) )   {
-          double x1 = params[0];
-          double x2 = params[1];
-          double y1 = params[2];
-          double y2 = params[3];
-          double z  = params[4];
-          double r  = params[5];
-          bool   atMinusZ = params[6] > 0.5;
-          double x            = atMinusZ ? x1 : x2;
-          double h            = 0;
-          bool   intersec     = false; // union or intersection solid
-          double displacement = 0;
-          double startPhi     = 0;
-          double halfZ        = z;
-          double halfOpeningAngle = std::asin( x / std::abs( r ))/units::deg;
-          /// calculate the displacement of the tubs w.r.t. to the trap, determine the opening angle of the tubs
-          double delta        = std::sqrt( r * r - x * x );
-
-#ifdef DIMENSION_DEBUG
-          cout << "setDimensions: [" << PSEUDOTRAP_TAG << "]" << endl
-               << right_matrix->GetTitle() << endl;
-#endif
-          // Implementation from :
-          // https://cmssdt.cern.ch/lxr/source/Fireworks/Geometry/src/TGeoMgrFromDdd.cc#0538
-          if( r < 0 && std::abs(r) >= x )  {
-            intersec = true;       // intersection solid
-            h = y1 < y2 ? y2 : y1; // tubs half height
-            h += h/20.;            // enlarge a bit - for subtraction solid
-            if( atMinusZ )    {
-              displacement = - halfZ - delta; 
-              startPhi     =  90.0 - halfOpeningAngle;
-            }
-            else    {
-              displacement =   halfZ + delta;
-              startPhi     = -90.0 - halfOpeningAngle;
-            }
-          }
-          else if( r > 0 && std::abs(r) >= x )  {
-            if( atMinusZ )    {
-              displacement = - halfZ + delta;
-              startPhi     = 270.0 - halfOpeningAngle;
-              h = y1;
-            }
-            else
-            {
-              displacement =   halfZ - delta; 
-              startPhi     =  90.0 - halfOpeningAngle;
-              h = y2;
-            }    
-          }
-          else  {
-            except(PSEUDOTRAP_TAG,"Check parameters of the PseudoTrap!");   
-          }
-          double trd2_dim[] = { x1, x2, y1, y2, halfZ };
-          double tube_dim[] = { 0.0, std::abs(r), h, startPhi, startPhi + halfOpeningAngle*2. };
-          if( intersec && left_solid->IsA() == TGeoTrd2::Class() )   {
-            left_solid->SetDimensions(trd2_dim);
-            right_solid->SetDimensions(tube_dim);
-          }
-          else if ( right_solid->IsA() == TGeoCompositeShape::Class() )   {
-            double box_dim[] = { 1.1*x, 1.1*h, std::sqrt(r*r-x*x) };
-            TGeoCompositeShape* comp = (TGeoCompositeShape*)right_solid;
-            TGeoBoolNode* b_s = comp->GetBoolNode();
-            b_s->GetLeftShape()->SetDimensions(tube_dim);
-            b_s->GetRightShape()->SetDimensions(box_dim);
-            left_solid->SetDimensions(trd2_dim);
-          }
-          else  {
-            except("PseudoTrap","+++ Incompatible change of parameters.");
-          }
-          ((TGeoTranslation*)right_matrix)->SetTranslation(0,0,displacement);
-          stringstream str;
-          str << x1 << " " << x2 << " " << y1 << " " << y2 << " " << z << " "
-              << r << " " << char(atMinusZ ? '1' : '0') << " ";
-          right_matrix->SetTitle(str.str().c_str());
-          return;
-        }
-        // In general TGeoCompositeShape instances have an empty SetDimension call
-        else if ( isInstance<SubtractionSolid>(solid) )   {
-        }
-        else if ( isInstance<UnionSolid>(solid) )   {
-        }
-        else if ( isInstance<IntersectionSolid>(solid) )   {
-        }
-#ifdef DIMENSION_DEBUG
-        throw runtime_error("Composite shape. setDimensions is not implemented!");
-#endif
-      }
-      else  {
-        printout(ERROR,"Solid","Failed to set dimensions for shape of type:%s.",
-                 cl->GetName());
-      }
-      return;
-    }
-    except("Solid","set_shape_dimensions: Failed to set dimensions [Invalid handle].");
-  }
-  /// Set the shape dimensions. As for the TGeo shape, but angles in rad rather than degrees.
-  void set_shape_dimensions(Solid solid, const vector<double>& params)   {
-    set_shape_dimensions(solid.ptr(), params);
-  }
-}
-
-/// Pretty print of solid attributes
-string dd4hep::toStringSolid(const TGeoShape* shape, int precision)   {
-  stringstream log;
-
-  if ( !shape )  {
-    return "[Invalid shape]";
-  }
-
-  TClass* cl = shape->IsA();
-  int prec = log.precision();
-  log.setf(ios::fixed,ios::floatfield);
-  log << setprecision(precision);
-  log << cl->GetName();
-  if ( cl == TGeoBBox::Class() )   {
-    TGeoBBox* sh = (TGeoBBox*) shape;
-    log << " x:" << sh->GetDX()
-        << " y:" << sh->GetDY()
-        << " z:" << sh->GetDZ();
-  }
-  else if (cl == TGeoHalfSpace::Class()) {
-    TGeoHalfSpace* sh = (TGeoHalfSpace*)(const_cast<TGeoShape*>(shape));
-    log << " Point:  (" << sh->GetPoint()[0] << ", " << sh->GetPoint()[1] << ", " << sh->GetPoint()[2] << ") " 
-        << " Normal: (" << sh->GetNorm()[0]  << ", " << sh->GetNorm()[1]  << ", " << sh->GetNorm()[2]  << ") ";
-  }
-  else if (cl == TGeoTube::Class()) {
-    const TGeoTube* sh = (const TGeoTube*) shape;
-    log << " rmin:" << sh->GetRmin() << " rmax:" << sh->GetRmax() << " dz:" << sh->GetDz();
-  }
-  else if (cl == TGeoTubeSeg::Class()) {
-    const TGeoTubeSeg* sh = (const TGeoTubeSeg*) shape;
-    log << " rmin:" << sh->GetRmin() << " rmax:" << sh->GetRmax() << " dz:" << sh->GetDz()
-        << " Phi1:" << sh->GetPhi1() << " Phi2:" << sh->GetPhi2();
-  }
-  else if (cl == TGeoCtub::Class()) {
-    const TGeoCtub* sh = (const TGeoCtub*) shape;
-    const Double_t*	hi = sh->GetNhigh();
-    const Double_t*	lo = sh->GetNlow();
-    log << " rmin:" << sh->GetRmin() << " rmax:" << sh->GetRmax() << " dz:" << sh->GetDz()
-        << " Phi1:" << sh->GetPhi1() << " Phi2:" << sh->GetPhi2()
-        << " lx:" << lo[0] << " ly:" << lo[1] << " lz:" << lo[2]
-        << " hx:" << hi[0] << " hy:" << hi[1] << " hz:" << hi[2];
-  }
-  else if (cl == TGeoEltu::Class()) {
-    const TGeoEltu* sh = (const TGeoEltu*) shape;
-    log << " A:" << sh->GetA() << " B:" << sh->GetB() << " dz:" << sh->GetDz();
-  }
-  else if (cl == TGeoTrd1::Class()) {
-    const TGeoTrd1* sh = (const TGeoTrd1*) shape;
-    log << " x1:" << sh->GetDx1() << " x2:" << sh->GetDx2() << " y:" << sh->GetDy() << " z:" << sh->GetDz();
-  }
-  else if (cl == TGeoTrd2::Class()) {
-    const TGeoTrd2* sh = (const TGeoTrd2*) shape;
-    log << " x1:" << sh->GetDx1() << " x2:" << sh->GetDx2()
-        << " y1:" << sh->GetDy1() << " y2:" << sh->GetDy2() << " z:" << sh->GetDz();
-  }
-  else if (cl == TGeoTrap::Class()) {
-    const TGeoTrap* sh = (const TGeoTrap*) shape;
-    log << " dz:" << sh->GetDz() << " Theta:" << sh->GetTheta() << " Phi:" << sh->GetPhi()
-        << " H1:" << sh->GetH1() << " Bl1:"   << sh->GetBl1()   << " Tl1:" << sh->GetTl1() << " Alpha1:" << sh->GetAlpha1()
-        << " H2:" << sh->GetH2() << " Bl2:"   << sh->GetBl2()   << " Tl2:" << sh->GetTl2() << " Alpha2:" << sh->GetAlpha2();
-  }
-  else if (cl == TGeoHype::Class()) {
-    const TGeoHype* sh = (const TGeoHype*) shape;
-    log << " rmin:" << sh->GetRmin() << " rmax:"  << sh->GetRmax() << " dz:" << sh->GetDz()
-        << " StIn:" << sh->GetStIn() << " StOut:" << sh->GetStOut();
-  }
-  else if (cl == TGeoPgon::Class()) {
-    const TGeoPgon* sh = (const TGeoPgon*) shape;
-    log << " Phi1:"   << sh->GetPhi1()   << " dPhi:" << sh->GetDphi()
-        << " NEdges:" << sh->GetNedges() << " Nz:" << sh->GetNz();
-    for(int i=0, n=sh->GetNz(); i<n; ++i)  {
-      log << " i=" << i << " z:" << sh->GetZ(i)
-          << " r:[" << sh->GetRmin(i) << "," << sh->GetRmax(i) << "]";
-    }
-  }
-  else if (cl == TGeoPcon::Class()) {
-    const TGeoPcon* sh = (const TGeoPcon*) shape;
-    log << " Phi1:" << sh->GetPhi1() << " dPhi:" << sh->GetDphi() << " Nz:" << sh->GetNz();
-    for(int i=0, n=sh->GetNz(); i<n; ++i)  {
-      log << " i=" << i << " z:" << sh->GetZ(i)
-          << " r:[" << sh->GetRmin(i) << "," << sh->GetRmax(i) << "]";
-    }
-  }
-  else if (cl == TGeoCone::Class()) {
-    const TGeoCone* sh = (const TGeoCone*) shape;
-    log << " rmin1:" << sh->GetRmin1() << " rmax1:" << sh->GetRmax1()
-        << " rmin2:" << sh->GetRmin2() << " rmax2:" << sh->GetRmax2()
-        << " dz:"    << sh->GetDz();
-  }
-  else if (cl == TGeoConeSeg::Class()) {
-    const TGeoConeSeg* sh = (const TGeoConeSeg*) shape;
-    log << " rmin1:" << sh->GetRmin1() << " rmax1:" << sh->GetRmax1()
-        << " rmin2:" << sh->GetRmin2() << " rmax2:" << sh->GetRmax2()
-        << " dz:"    << sh->GetDz()
-        << " Phi1:"  << sh->GetPhi1() << " Phi2:" << sh->GetPhi2();
-  }
-  else if (cl == TGeoParaboloid::Class()) {
-    const TGeoParaboloid* sh = (const TGeoParaboloid*) shape;
-    log << " dz:" << sh->GetDz() << " RLo:" << sh->GetRlo() << " Rhi:" << sh->GetRhi();
-  }
-  else if (cl == TGeoSphere::Class()) {
-    const TGeoSphere* sh = (const TGeoSphere*) shape;
-    log << " rmin:" << sh->GetRmin() << " rmax:" << sh->GetRmax()
-        << " Phi1:" << sh->GetPhi1() << " Phi2:" << sh->GetPhi2()
-        << " Theta1:" << sh->GetTheta1() << " Theta2:" << sh->GetTheta2();
-  }
-  else if (cl == TGeoTorus::Class()) {
-    const TGeoTorus* sh = (const TGeoTorus*) shape;
-    log << " rmin:" << sh->GetRmin() << " rmax:" << sh->GetRmax() << " r:" << sh->GetR()
-        << " Phi1:" << sh->GetPhi1() << " dPhi:" << sh->GetDphi();
-  }
-  else if (cl == TGeoArb8::Class()) {
-    TGeoArb8* sh = (TGeoArb8*) shape;
-    const Double_t* v = sh->GetVertices();
-    log << " dz:" << sh->GetDz();
-    for(int i=0; i<8; ++i) {
-      log << " x[" << i << "]:" << *v; ++v;
-      log << " y[" << i << "]:" << *v; ++v;
-    }
-  }
-  else if (cl == TGeoXtru::Class()) {
-    const TGeoXtru* sh = (const TGeoXtru*) shape;
-    log << " X:   " << sh->GetX(0)
-        << " Y:   " << sh->GetY(0)
-        << " Z:   " << sh->GetZ(0)
-        << " Xoff:" << sh->GetXOffset(0)
-        << " Yoff:" << sh->GetYOffset(0)
-        << " Nvtx:" << sh->GetNvert()
-        << " Nz:  " << sh->GetNz();
-  }
-  else if (shape->IsA() == TGeoCompositeShape::Class()) {
-    const TGeoCompositeShape* sh = (const TGeoCompositeShape*) shape;
-    const TGeoBoolNode* boolean = sh->GetBoolNode();
-    const TGeoShape* left  = boolean->GetLeftShape();
-    const TGeoShape* right = boolean->GetRightShape();
-    TGeoBoolNode::EGeoBoolType oper = boolean->GetBooleanOperator();
-    if (oper == TGeoBoolNode::kGeoSubtraction)
-      log << "Subtraction: ";
-    else if (oper == TGeoBoolNode::kGeoUnion)
-      log << "Union: ";
-    else if (oper == TGeoBoolNode::kGeoIntersection)
-      log << "Intersection: ";
-    log << " Left:" << toStringSolid(left) << " Right:" << toStringSolid(right);
-  }
-  log << setprecision(prec);
-  return log.str();
-}
-
-/// Output mesh vertices to string
-string dd4hep::toStringMesh(const TGeoShape* shape, int prec)  {
-  Solid sol(shape);
-  stringstream os;
-
-  // Prints shape parameters
-  Int_t nvert = 0, nsegs = 0, npols = 0;
-
-  sol->GetMeshNumbers(nvert, nsegs, npols);
-  Double_t* points = new Double_t[3*nvert];
-  sol->SetPoints(points);
-
-  os << setw(16) << left << sol->IsA()->GetName()
-     << " " << nvert << " Mesh-points:" << endl;
-  os << setw(16) << left << sol->IsA()->GetName() << " " << sol->GetName()
-     << " N(mesh)=" << sol->GetNmeshVertices()
-     << "  N(vert)=" << nvert << "  N(seg)=" << nsegs << "  N(pols)=" << npols << endl;
-    
-  for(Int_t i=0; i<nvert; ++i)   {
-    Double_t* p = points + 3*i;
-    os << setw(16) << left << sol->IsA()->GetName() << " " << setw(3) << left << i
-       << " Local  ("  << setw(7) << setprecision(prec) << fixed << right << p[0]
-       << ", "         << setw(7) << setprecision(prec) << fixed << right << p[1]
-       << ", "         << setw(7) << setprecision(prec) << fixed << right << p[2]
-       << ")" << endl;
-  }
-  Box box = sol;
-  const Double_t* org = box->GetOrigin();
-  os << setw(16) << left << sol->IsA()->GetName()
-     << " Bounding box: "
-     << " dx="        << setw(7) << setprecision(prec) << fixed << right << box->GetDX()
-     << " dy="        << setw(7) << setprecision(prec) << fixed << right << box->GetDY()
-     << " dz="        << setw(7) << setprecision(prec) << fixed << right << box->GetDZ()
-     << " Origin: x=" << setw(7) << setprecision(prec) << fixed << right << org[0]
-     << " y="         << setw(7) << setprecision(prec) << fixed << right << org[1]
-     << " z="         << setw(7) << setprecision(prec) << fixed << right << org[2]
-     << endl;
-  
-  /// -------------------- DONE --------------------
-  delete [] points;
-  return os.str();
-}
-
 template <typename T> void Solid_type<T>::_setDimensions(double* param)   const {
-  this->ptr()->SetDimensions(param);
-  this->ptr()->ComputeBBox();
+  auto p = this->access(); // Ensure we have a valid handle!
+  p->SetDimensions(param);
+  p->ComputeBBox();
 }
 
 /// Assign pointrs and register solid to geometry
@@ -1080,11 +101,11 @@ Solid_type<T>::divide(const Volume& voldiv, const string& divname,
 
 /// Constructor to create an anonymous new box object (retrieves name from volume)
 ShapelessSolid::ShapelessSolid(const string& nam)  {
-  _assign(new TGeoShapeAssembly(), nam, "Assembly", true);
+  _assign(new TGeoShapeAssembly(), nam, SHAPELESS_TAG, true);
 }
 
 void Box::make(const string& nam, double x_val, double y_val, double z_val)   {
-  _assign(new TGeoBBox(nam.c_str(), x_val, y_val, z_val), "", "Box", true);
+  _assign(new TGeoBBox(nam.c_str(), x_val, y_val, z_val), "", BOX_TAG, true);
 }
 
 /// Set the box dimensionsy
@@ -1111,12 +132,12 @@ double Box::z() const {
 
 /// Internal helper method to support object construction
 void HalfSpace::make(const string& nam, const double* const point, const double* const normal)   {
-  _assign(new TGeoHalfSpace(nam.c_str(),(Double_t*)point, (Double_t*)normal), "", "HalfSpace",true);
+  _assign(new TGeoHalfSpace(nam.c_str(),(Double_t*)point, (Double_t*)normal), "", HALFSPACE_TAG,true);
 }
 
 /// Constructor to be used when creating a new object
 Polycone::Polycone(double startPhi, double deltaPhi) {
-  _assign(new TGeoPcon(startPhi/units::deg, deltaPhi/units::deg, 0), "", "Polycone", false);
+  _assign(new TGeoPcon(startPhi/units::deg, deltaPhi/units::deg, 0), "", POLYCONE_TAG, false);
 }
 
 /// Constructor to be used when creating a new polycone object. Add at the same time all Z planes
@@ -1137,7 +158,7 @@ Polycone::Polycone(double startPhi, double deltaPhi,
     params.emplace_back(rmin[i] );
     params.emplace_back(rmax[i] );
   }
-  _assign(new TGeoPcon(&params[0]), "", "Polycone", true);
+  _assign(new TGeoPcon(&params[0]), "", POLYCONE_TAG, true);
 }
 
 /// Constructor to be used when creating a new polycone object. Add at the same time all Z planes
@@ -1157,12 +178,12 @@ Polycone::Polycone(double startPhi, double deltaPhi, const vector<double>& r, co
     params.emplace_back(0.0  );
     params.emplace_back(r[i] );
   }
-  _assign(new TGeoPcon(&params[0]), "", "Polycone", true);
+  _assign(new TGeoPcon(&params[0]), "", POLYCONE_TAG, true);
 }
 
 /// Constructor to be used when creating a new object
 Polycone::Polycone(const string& nam, double startPhi, double deltaPhi) {
-  _assign(new TGeoPcon(nam.c_str(), startPhi/units::deg, deltaPhi/units::deg, 0), "", "Polycone", false);
+  _assign(new TGeoPcon(nam.c_str(), startPhi/units::deg, deltaPhi/units::deg, 0), "", POLYCONE_TAG, false);
 }
 
 /// Constructor to be used when creating a new polycone object. Add at the same time all Z planes
@@ -1183,7 +204,7 @@ Polycone::Polycone(const string& nam, double startPhi, double deltaPhi,
     params.emplace_back(rmin[i] );
     params.emplace_back(rmax[i] );
   }
-  _assign(new TGeoPcon(&params[0]), nam, "Polycone", true);
+  _assign(new TGeoPcon(&params[0]), nam, POLYCONE_TAG, true);
 }
 
 /// Constructor to be used when creating a new polycone object. Add at the same time all Z planes
@@ -1203,7 +224,7 @@ Polycone::Polycone(const string& nam, double startPhi, double deltaPhi, const ve
     params.emplace_back(0.0  );
     params.emplace_back(r[i] );
   }
-  _assign(new TGeoPcon(&params[0]), nam, "Polycone", true);
+  _assign(new TGeoPcon(&params[0]), nam, POLYCONE_TAG, true);
 }
 
 /// Add Z-planes to the Polycone
@@ -1212,7 +233,7 @@ void Polycone::addZPlanes(const vector<double>& rmin, const vector<double>& rmax
   vector<double> params;
   size_t num = sh->GetNz();
   if (rmin.size() < 2)   {
-    throw runtime_error("dd4hep: PolyCone::addZPlanes> Not enough Z planes. minimum is 2!");
+    except("PolyCone","++ addZPlanes: Not enough Z planes. minimum is 2!");
   }
   params.emplace_back(sh->GetPhi1());
   params.emplace_back(sh->GetDphi());
@@ -1237,7 +258,7 @@ ConeSegment::ConeSegment(double dz,
                          double startPhi,  double endPhi)
 {
   _assign(new TGeoConeSeg(dz, rmin1, rmax1, rmin2, rmax2,
-                          startPhi/units::deg, endPhi/units::deg), "", "ConeSegment", true);
+                          startPhi/units::deg, endPhi/units::deg), "", CONESEGMENT_TAG, true);
 }
 
 /// Constructor to be used when creating a new cone segment object
@@ -1248,7 +269,7 @@ ConeSegment::ConeSegment(const string& nam,
                          double startPhi,  double endPhi)
 {
   _assign(new TGeoConeSeg(nam.c_str(), dz, rmin1, rmax1, rmin2, rmax2,
-                          startPhi/units::deg, endPhi/units::deg), "", "ConeSegment", true);
+                          startPhi/units::deg, endPhi/units::deg), "", CONESEGMENT_TAG, true);
 }
 
 /// Set the cone segment dimensions
@@ -1263,7 +284,7 @@ ConeSegment& ConeSegment::setDimensions(double dz,
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Cone::make(const string& nam, double z, double rmin1, double rmax1, double rmin2, double rmax2) {
-  _assign(new TGeoCone(nam.c_str(), z, rmin1, rmax1, rmin2, rmax2 ), "", "Cone", true);
+  _assign(new TGeoCone(nam.c_str(), z, rmin1, rmax1, rmin2, rmax2 ), "", CONE_TAG, true);
 }
 
 /// Set the box dimensions (startPhi=0.0, endPhi=2*pi)
@@ -1275,7 +296,7 @@ Cone& Cone::setDimensions(double z, double rmin1, double rmax1, double rmin2, do
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Tube::make(const string& nam, double rmin, double rmax, double z, double startPhi, double endPhi) {
-  _assign(new TGeoTubeSeg(nam.c_str(), rmin, rmax, z, startPhi/units::deg, endPhi/units::deg),nam,"Tube",true);
+  _assign(new TGeoTubeSeg(nam.c_str(), rmin, rmax, z, startPhi/units::deg, endPhi/units::deg),nam,TUBE_TAG,true);
 }
 
 /// Set the tube dimensions
@@ -1301,7 +322,7 @@ CutTube::CutTube(const string& nam,
 /// Constructor to be used when creating a new object with attribute initialization
 void CutTube::make(const string& nam, double rmin, double rmax, double dz, double startPhi, double endPhi,
                    double lx, double ly, double lz, double tx, double ty, double tz)  {
-  _assign(new TGeoCtub(nam.c_str(), rmin,rmax,dz,startPhi,endPhi,lx,ly,lz,tx,ty,tz),"","CutTube",true);
+  _assign(new TGeoCtub(nam.c_str(), rmin,rmax,dz,startPhi,endPhi,lx,ly,lz,tx,ty,tz),"",CUTTUBE_TAG,true);
 }
 
 /// Constructor to create a truncated tube object with attribute initialization
@@ -1412,12 +433,12 @@ void TruncatedTube::make(const string& nam,
 
 /// Constructor to be used when creating a new object with attribute initialization
 void EllipticalTube::make(const string& nam, double a, double b, double dz) {
-  _assign(new TGeoEltu(nam.c_str(), a, b, dz), "", "EllipticalTube", true);
+  _assign(new TGeoEltu(nam.c_str(), a, b, dz), "", ELLIPTICALTUBE_TAG, true);
 }
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Trd1::make(const string& nam, double x1, double x2, double y, double z) {
-  _assign(new TGeoTrd1(nam.c_str(), x1, x2, y, z ), "", "Trd1", true);
+  _assign(new TGeoTrd1(nam.c_str(), x1, x2, y, z ), "", TRD1_TAG, true);
 }
 
 /// Set the Trd1 dimensions
@@ -1429,7 +450,7 @@ Trd1& Trd1::setDimensions(double x1, double x2, double y, double z) {
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Trd2::make(const string& nam, double x1, double x2, double y1, double y2, double z) {
-  _assign(new TGeoTrd2(nam.c_str(), x1, x2, y1, y2, z ), "", "Trd2", true);
+  _assign(new TGeoTrd2(nam.c_str(), x1, x2, y1, y2, z ), "", TRD2_TAG, true);
 }
 
 /// Set the Trd2 dimensions
@@ -1441,7 +462,7 @@ Trd2& Trd2::setDimensions(double x1, double x2, double y1, double y2, double z) 
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Paraboloid::make(const string& nam, double r_low, double r_high, double delta_z) {
-  _assign(new TGeoParaboloid(nam.c_str(), r_low, r_high, delta_z ), "", "Paraboloid", true);
+  _assign(new TGeoParaboloid(nam.c_str(), r_low, r_high, delta_z ), "", PARABOLOID_TAG, true);
 }
 
 /// Set the Paraboloid dimensions
@@ -1453,7 +474,7 @@ Paraboloid& Paraboloid::setDimensions(double r_low, double r_high, double delta_
 
 /// Constructor to create a new anonymous object with attribute initialization
 void Hyperboloid::make(const string& nam, double rin, double stin, double rout, double stout, double dz) {
-  _assign(new TGeoHype(nam.c_str(), rin, stin/units::deg, rout, stout/units::deg, dz), "", "Hyperboloid", true);
+  _assign(new TGeoHype(nam.c_str(), rin, stin/units::deg, rout, stout/units::deg, dz), "", HYPERBOLOID_TAG, true);
 }
 
 /// Set the Hyperboloid dimensions
@@ -1467,7 +488,7 @@ Hyperboloid& Hyperboloid::setDimensions(double rin, double stin, double rout, do
 void Sphere::make(const string& nam, double rmin, double rmax, double startTheta, double endTheta, double startPhi, double endPhi) {
   _assign(new TGeoSphere(nam.c_str(), rmin, rmax,
                          startTheta/units::deg, endTheta/units::deg,
-                         startPhi/units::deg,   endPhi/units::deg), "", "Sphere", true);
+                         startPhi/units::deg,   endPhi/units::deg), "", SPHERE_TAG, true);
 }
 
 /// Set the Sphere dimensions
@@ -1479,7 +500,7 @@ Sphere& Sphere::setDimensions(double rmin, double rmax, double startTheta, doubl
 
 /// Constructor to be used when creating a new object with attribute initialization
 void Torus::make(const string& nam, double r, double rmin, double rmax, double startPhi, double deltaPhi) {
-  _assign(new TGeoTorus(nam.c_str(), r, rmin, rmax, startPhi/units::deg, deltaPhi/units::deg), "", "Torus", true);
+  _assign(new TGeoTorus(nam.c_str(), r, rmin, rmax, startPhi/units::deg, deltaPhi/units::deg), "", TORUS_TAG, true);
 }
 
 /// Set the Torus dimensions
@@ -1495,7 +516,7 @@ Trap::Trap(double z, double theta, double phi,
            double h2, double bl2, double tl2, double alpha2) {
   _assign(new TGeoTrap(z, theta/units::deg, phi/units::deg,
                        h1, bl1, tl1, alpha1/units::deg,
-                       h2, bl2, tl2, alpha2/units::deg), "", "Trap", true);
+                       h2, bl2, tl2, alpha2/units::deg), "", TRAP_TAG, true);
 }
 
 /// Constructor to be used when creating a new anonymous object with attribute initialization
@@ -1505,7 +526,7 @@ Trap::Trap(const string& nam,
            double h2, double bl2, double tl2, double alpha2) {
   _assign(new TGeoTrap(nam.c_str(), z, theta/units::deg, phi/units::deg,
                        h1, bl1, tl1, alpha1/units::deg,
-                       h2, bl2, tl2, alpha2/units::deg), "", "Trap", true);
+                       h2, bl2, tl2, alpha2/units::deg), "", TRAP_TAG, true);
 }
 
 /// Constructor to be used when creating a new anonymous object with attribute initialization
@@ -1519,7 +540,7 @@ void Trap::make(const string& nam, double pz, double py, double px, double pLTX)
   double alpha1 = (pLTX - px) / py;
   _assign(new TGeoTrap(nam.c_str(), z, theta, phi,
                        h, bl, tl, alpha1/units::deg,
-                       h, bl, tl, alpha1/units::deg), "", "Trap", true);
+                       h, bl, tl, alpha1/units::deg), "", TRAP_TAG, true);
 }
 
 /// Set the trap dimensions
@@ -1637,7 +658,7 @@ void PolyhedraRegular::make(const string& nam, int nsides, double rmin, double r
   else if (rmax < 0e0)
     throw runtime_error("dd4hep: PolyhedraRegular: Illegal argument rmax:<" + _toString(rmax) + "> is invalid!");
   double params[] = { start/units::deg, delta/units::deg, double(nsides), 2e0, zpos, rmin, rmax, zneg, rmin, rmax };
-  _assign(new TGeoPgon(params), nam, "Polyhedra", false);
+  _assign(new TGeoPgon(params), nam, POLYHEDRA_TAG, false);
   //_setDimensions(&params[0]);
 }
 
@@ -1661,7 +682,7 @@ void Polyhedra::make(const string& nam, int nsides, double start, double delta,
     temp.emplace_back(rmin[i]);
     temp.emplace_back(rmax[i]);
   }
-  _assign(new TGeoPgon(&temp[0]), nam, "Polyhedra", false);
+  _assign(new TGeoPgon(&temp[0]), nam, POLYHEDRA_TAG, false);
 }
 
 /// Helper function to create the polyhedron
@@ -1674,7 +695,7 @@ void ExtrudedPolygon::make(const string& nam,
                            const vector<double>& sec_scale)
 {
   TGeoXtru* solid = new TGeoXtru(sec_z.size());
-  _assign(solid, nam, "ExtrudedPolygon", false);
+  _assign(solid, nam, EXTRUDEDPOLYGON_TAG, false);
   // No need to transform coordinates to cm. We are in the dd4hep world: all is already in cm.
   solid->DefinePolygon(pt_x.size(), &(*pt_x.begin()), &(*pt_y.begin()));
   for( size_t i = 0; i < sec_z.size(); ++i )
@@ -1683,7 +704,7 @@ void ExtrudedPolygon::make(const string& nam,
 
 /// Creator method
 void EightPointSolid::make(const string& nam, double dz, const double* vtx)   {
-  _assign(new TGeoArb8(nam.c_str(), dz, (double*)vtx), "", "EightPointSolid", true);
+  _assign(new TGeoArb8(nam.c_str(), dz, (double*)vtx), "", EIGHTPOINTSOLID_TAG, true);
 }
 
 /// Constructor to be used when creating a new object. Position is identity, Rotation is the identity rotation
@@ -1883,6 +904,7 @@ INSTANTIATE(TGeoTube);
 INSTANTIATE(TGeoTubeSeg);
 INSTANTIATE(TGeoEltu);
 INSTANTIATE(TGeoXtru);
+INSTANTIATE(TGeoGtra);
 INSTANTIATE(TGeoHype);
 INSTANTIATE(TGeoTrap);
 INSTANTIATE(TGeoTrd1);
